pipeline RNA-seq single cells
In the main directory store the R-project file, a csv file with a list of genes you wish to analyze, and a folder named ‘.01. Transcript Counts’. In the latter you will place one folder per condition with .tsv files of transcript counts.

Install required packages. 

In order for the pipeline to work, you will also need to source RacdeID3_StemID2 which is created by Dominik Grun.